switched denorm table to OOP, some more abstracting
