+++ app/api/denormalized_user_challenge/__init__.py
@@ -1,0 +1,1 @@
+from denormalized_user_challenge_repository import Denormalized_User_Challenge_Repository
