+++ app/api/user_challenge/__init__.py
@@ -1,1 +1,2 @@
 # import the errors
+from app.api.user_challenge.User_Challenge_Repository import User_Challenge_Repository
