--- app/api/user_challenge_dimension_table/__init__.py
@@ -1,1 +1,0 @@
-from user_challenge_dimension_repository import create_dimension_table_entry, get_entries
